native: MainLoopStep runs a complete frame
Five frames, exit 0. The engine's own per-frame loop - BeginFrame,
GameLogicStep, GameStateMachine, GameTick, DrawScene, PresentFrame - now runs
to completion natively. The frame is still black: nothing has been staged into
the scene graph yet, and two callbacks are not linked. But it no longer dies.

Three defects stood between here and there, and the last one is in the memory
model itself.

1. 276 VAs were cast straight to host pointers, bypassing the seam
   (`*(int *)(iVar2 + 0xd8)` where iVar2 holds a VA). tools/decomp/audit_va_deref.py
   finds them from clang's OWN diagnostics rather than a regex: -Wint-to-pointer-cast
   fires on exactly "cast to a pointer from a SMALLER integer type", and a
   32-bit value used as an address in this engine is a VA by construction. The
   file:line:col it reports is precise enough to rewrite from. 276 sites in 93
   files; the 77 FUNCTION-pointer casts are reported and left alone, because a
   stored code VA needs MK4_ResolveCode, not the data seam - which one applies
   is a per-site judgement.

   Those casts were invisible because the native build passes
   -Wno-int-to-pointer-cast. 381 diagnostics were being suppressed.

2. MK4_ResolveCode passed unknown VAs through unchanged, so an unconverted
   callback became a JUMP to a raw VA and the process died with nothing to say.
   It now reports the VA once and returns a no-op, so one missing target does
   not hide the rest of the frame. That immediately named both:
   0x45cda0/0x45cdb0, in pending_match_variants.c.

3. MK4_PTR computed its offset UNSIGNED. A base-0 packed table is spelled
   MK4_VA(T, 0) and indexed by the VA itself, so the base is deliberately below
   the image and the index adds the VA back - but `0 - 0x400000` unsigned is
   0xffc00000, and adding that to a 64-bit pointer does not wrap back. The
   access landed exactly 4 GB high, which is what Helper_TickAlt was dying on.
   Signed, the base is arena - 4 MB and the index brings it home. In-range VAs
   are unaffected, which is why every byte-identical gate still holds.

The range-check build had to be corrected for the same reason: it was
substituting the arena base for out-of-range VAs, which made fifteen correct
base-0 accesses look like faults AND let the checked build finish a frame the
real one could not. A diagnostic that returns a different address is not
diagnosing the program any more. It now always returns what the plain macro
computes and only reports.

Gates: matching byte-identical; native-char-check byte-identical to the
verified bundle; native-frame-check 62300 px; both wasm32 parity gates
byte-identical; verify_emit 14/14, verify_mesh 10/10, verify_project 30/30,
verify_rot 18/18.

+++ include/portable/mem_model.h
@@ -61,8 +61,16 @@ extern unsigned char *g_mk4Arena;       /* base of the reserved data arena */
 void *MK4_PtrChecked(unsigned va, const char *file, int line);
 #define MK4_PTR(va)  MK4_PtrChecked((unsigned)(va), __FILE__, __LINE__)
 #else
+/* The offset is SIGNED, and that is load-bearing rather than tidiness.
+ *
+ * A BASE-0 packed table is spelled MK4_VA(T, 0) and indexed by the VA itself -
+ * g_siblingTable[idx] with idx = VA/4 - so the base is deliberately below the
+ * image and the index adds the VA back. Computed unsigned, `0 - 0x400000` is
+ * 0xffc00000, and adding THAT to a 64-bit pointer does not wrap back: the
+ * access lands 4 GB high and dies. Signed, the base is arena - 4 MB and the
+ * index brings it home. In-range VAs are unaffected either way. */
 #define MK4_PTR(va) \
-    ((void *)(g_mk4Arena + ((unsigned)(va) - MK4_ORIG_IMAGE_BASE)))
+    ((void *)(g_mk4Arena + ((long)(unsigned)(va) - (long)MK4_ORIG_IMAGE_BASE)))
 #endif
 #define MK4_NODE(T, idx)     ((T *)MK4_PTR((unsigned)(idx) * 4u))
 #define MK4_VA(T, va)        ((T *)MK4_PTR((unsigned)(va)))
